# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/galaxies": {
      "get": {
        "tags": [
          "routes::galaxy"
        ],
        "operationId": "get_all_galaxies",
        "responses": {
          "200": {
            "$ref": "#/components/responses/GalaxiesList"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "routes::galaxy"
        ],
        "operationId": "create_galaxy",
        "requestBody": {
          "description": "data for creating the galaxy",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGalaxyData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GalaxyCreated"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/galaxies/{galaxy_id}": {
      "get": {
        "tags": [
          "routes::galaxy"
        ],
        "operationId": "get_galaxy",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SpecificGalaxy"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "routes::galaxy"
        ],
        "operationId": "update_galaxy",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "data for updating the galaxy",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGalaxyData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GalaxyUpdated"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "routes::galaxy"
        ],
        "operationId": "delete_galaxy",
        "parameters": [
          {
            "name": "galaxy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GalaxyDeleted"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "routes::auth"
        ],
        "operationId": "login",
        "requestBody": {
          "description": "login data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthResponse"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        },
        "security": [
          {}
        ]
      }
    },
    "/logout": {
      "delete": {
        "tags": [
          "routes::auth"
        ],
        "operationId": "logout",
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "routes::auth"
        ],
        "operationId": "register",
        "requestBody": {
          "description": "registration data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AuthResponse"
          },
          "400": {
            "$ref": "#/components/responses/ValidationResponse"
          },
          "409": {
            "$ref": "#/components/responses/AlreadyExistsResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        },
        "security": [
          {}
        ]
      }
    },
...
    "/verify": {
      "get": {
        "tags": [
          "routes::auth"
        ],
        "operationId": "verify",
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Credentials"
          },
          {
            "type": "object",
            "required": [
              "remember"
            ],
            "properties": {
              "remember": {
                "type": "boolean"
              }
            }
          }
        ]
      },
...
      "CreateGalaxyData": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "Credentials": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Password"
          },
          {
            "type": "object",
            "required": [
              "username"
            ],
            "properties": {
              "username": {
                "type": "string"
              }
            }
          }
        ]
      },
      "Galaxy": {
        "type": "object",
        "required": [
          "id",
          "name",
          "user_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Password": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
      "Planet": {
        "type": "object",
        "required": [
          "id",
          "name",
          "capacity",
          "galaxy_id"
        ],
        "properties": {
          "capacity": {
            "type": "integer",
            "format": "int32",
            "maximum": 2147483647,
            "minimum": 0
          },
          "galaxy_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "star_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "Star": {
        "type": "object",
        "required": [
          "id",
          "name",
          "nebula",
          "galaxy_id"
        ],
        "properties": {
          "galaxy_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "nebula": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "UpdateGalaxyData": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "minLength": 1
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "AlreadyExistsResponse": {
        "description": "The resource already exists",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "AuthResponse": {
        "description": "user authorized from session token",
        "headers": {
          "Set-Cookie": {
            "schema": {
              "type": "string"
            },
            "description": "The session ID is returned in a cookie named `session`. You need to include this cookie in subsequent requests."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "user"
              ],
              "properties": {
                "user": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "GalaxiesList": {
        "description": "all user galaxies",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Galaxy"
              }
            }
          }
        }
      },
      "GalaxyCreated": {
        "description": "galaxy successfully created",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Galaxy"
            }
          }
        }
      },
      "GalaxyDeleted": {
        "description": "galaxy successfully deleted",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Galaxy"
            }
          }
        }
      },
      "GalaxyUpdated": {
        "description": "galaxy successfully updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Galaxy"
            }
          }
        }
      },
      "InternalErrorResponse": {
        "description": "An internal error occurred",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "Requested resources not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
...
      "SpecificGalaxy": {
        "description": "specific galaxy with all its stars and planets",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "galaxy",
                "stars",
                "planets"
              ],
              "properties": {
                "galaxy": {
                  "$ref": "#/components/schemas/Galaxy"
                },
                "planets": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Planet"
                  }
                },
                "stars": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Star"
                  }
                }
              }
            }
          }
        }
      },
...
      "UnauthorizeResponse": {
        "description": "User not authorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
...
      "ValidationResponse": {
        "description": "The body of the request contains incorrect data",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "session_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
  "security": [
    {
      "session_token": []
    }
  ]
}